requirements.txt
# Core Web Framework
streamlit==1.28.1
streamlit-folium==0.15.1
streamlit-authenticator==0.2.3

# Data Processing
pandas==2.1.4
numpy==1.26.4  # Compatible with Python 3.11
geopandas==0.14.0
shapely==2.0.2

# Visualization
plotly==5.18.0
folium==0.15.0
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning - LSTM & Time Series
tensorflow==2.20.0  # Better Python 3.11 support
scikit-learn==1.3.2
statsmodels==0.14.1

# Model Tracking & MLOps
mlflow==2.9.2
tensorboard==2.15.1

# API & Data Collection
requests==2.31.0
# openaq==2.0.0  # This version doesn't exist
python-openaq==1.1.0  # Correct package name
pyowm==3.3.0
python-dotenv==1.0.0

# Geocoding & Location Services
geopy==2.4.1
geocoder==1.38.1

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Notifications & Alerts
sendgrid==6.11.0
firebase-admin==6.3.0
twilio==8.11.1

# Task Scheduling (Optional - for production)
# apache-airflow==2.7.3  # Removed - complex dependencies, install separately if needed
celery==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2024.1
schedule==1.2.0
joblib==1.3.2  # Added - needed for model serialization

# Testing & Quality
pytest==7.4.3
black==23.12.1
flake8==7.0.0
